+++ stream.c
@@ -116,15 +116,15 @@ int rs_buffers_copy(rs_buffers_t *stream, int max_len)
     assert(len > 0);
 
     if ((unsigned) len > stream->avail_in) {
-        rs_trace("copy limited to %d available input bytes",
-                 stream->avail_in);
+        rs_trace("copy limited to %ld available input bytes",
+                 (long) stream->avail_in);
         len = stream->avail_in;
     }
 
 
     if ((unsigned) len > stream->avail_out) {
-        rs_trace("copy limited to %d available output bytes",
-                 stream->avail_out);
+        rs_trace("copy limited to %ld available output bytes",
+                 (long) stream->avail_out);
         len = stream->avail_out;
     }
